Move to middleware file rather than new folder
