+++ common/middleware.py
@@ -1,4 +1,5 @@
 from django.conf import settings
+from django.middleware.security import SecurityMiddleware
 from django.shortcuts import redirect
 from django.urls import reverse
 
@@ -15,4 +16,14 @@ class MaintenanceModeMiddleware:
             return redirect(reverse("maintenance"))
 
         response = self.get_response(request)
+        return response
+
+
+class CustomSecurityMiddleware(SecurityMiddleware):
+    """Extends the security middleware to include the X-Permitted-Cross-Domain-
+    Policies HTTP response header."""
+
+    def process_response(self, request, response):
+        response = super().process_response(request, response)
+        response["X-Permitted-Cross-Domain-Policies"] = "none"
         return response
